Config: Don't make oneMKL paths configurable
There are just too many oneMKL and compiler include and library paths for the
user to be able to specify them all. It's going to have to be automated to
some extent. So if we are going to be getting a oneMKL build2 package then I
assume that package will export the paths.

For example, these are some of the paths the oneMKL setvars script adds to
various environment variables (usually something ending with _PATH):

/opt/intel/oneapi/compiler/2025.2/bin
/opt/intel/oneapi/compiler/2025.2/lib
/opt/intel/oneapi/compiler/2025.2/opt/compiler/bin
/opt/intel/oneapi/compiler/2025.2/opt/compiler/lib
/opt/intel/oneapi/dpl/2022.9/include
/opt/intel/oneapi/mkl/2025.2/bin
/opt/intel/oneapi/mkl/2025.2/include
/opt/intel/oneapi/mkl/2025.2/lib
/opt/intel/oneapi/tbb/2022.2/env/../include
/opt/intel/oneapi/tbb/2022.2/env/../lib/intel64/gcc4.8
/opt/intel/oneapi/tcm/1.4/lib
/opt/intel/oneapi/umf/0.11/include
/opt/intel/oneapi/umf/0.11/lib